Add property "blockList" to ContentBlock
Decode "blocks" from RawContentState

Add base for tree example

Add base for component DraftEditorBlocks

Change blockList to blockMap in ContentBlock

Extend DraftOffsetKeyPath format

Switch to keep a flatten map of blocks in ContentState

Adapt splitBlockInContentState to split in the same parent
