+++ examples/tree/tree.html
@@ -1,0 +1,131 @@
+<!--
+Copyright (c) 2013-present, Facebook, Inc. All rights reserved.
+
+This file provided by Facebook is for non-commercial testing and evaluation
+purposes only. Facebook reserves all rights not expressly granted.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+FACEBOOK BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+-->
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title>Draft • Tree Editor</title>
+    <link rel="stylesheet" href="../../dist/Draft.css" />
+    <style>
+      .public-DraftStyleDefault-block {
+        padding-left: 20px;
+      }
+    </style>
+  </head>
+  <body>
+    <div id="target"></div>
+    <script src="../../node_modules/react/dist/react.js"></script>
+    <script src="../../node_modules/react-dom/dist/react-dom.js"></script>
+    <script src="../../node_modules/immutable/dist/immutable.js"></script>
+    <script src="../../node_modules/es6-shim/es6-shim.js"></script>
+    <script src="../../node_modules/babel-core/browser.js"></script>
+    <script src="../../dist/Draft.js"></script>
+    <script type="text/babel">
+      'use strict';
+
+      const {Editor, EditorState, ContentState, convertFromRaw} = Draft;
+
+      console.log('starting');
+      var baseContent = convertFromRaw({
+        blocks: [
+          {
+            type: 'unstyled',
+            text: 'Hello World',
+            blocks: [
+              {
+                type: 'unstyled',
+                text: 'Cool',
+                blocks: [
+                  {
+                    type: 'unstyled',
+                    text: 'Another deep block',
+                    inlineStyleRanges: [
+                      {
+                        offset: 8,
+                        length: 4,
+                        style: 'BOLD'
+                      }
+                    ]
+                  }
+                ]
+              },
+              {
+                type: 'unstyled',
+                text: 'Awesome'
+              }
+            ]
+          }
+        ],
+        entityMap: {}
+      });
+      console.log('baseContent is', baseContent.toJS());
+
+
+      class TreeEditorExample extends React.Component {
+        constructor(props) {
+          super(props);
+          this.state = {editorState: EditorState.createWithContent(baseContent)};
+
+          this.focus = () => this.refs.editor.focus();
+          this.onChange = (editorState) => this.setState({editorState});
+          this.logState = () => console.log(this.state.editorState.toJS());
+        }
+
+        render() {
+          return (
+            <div style={styles.root}>
+              <div style={styles.editor} onClick={this.focus}>
+                <Editor
+                  editorState={this.state.editorState}
+                  onChange={this.onChange}
+                  placeholder="Enter some text..."
+                  ref="editor"
+                />
+              </div>
+              <input
+                onClick={this.logState}
+                style={styles.button}
+                type="button"
+                value="Log State"
+              />
+            </div>
+          );
+        }
+      }
+
+      const styles = {
+        root: {
+          fontFamily: '\'Helvetica\', sans-serif',
+          padding: 20,
+          width: 600,
+        },
+        editor: {
+          border: '1px solid #ccc',
+          cursor: 'text',
+          minHeight: 80,
+          padding: 10,
+        },
+        button: {
+          marginTop: 10,
+          textAlign: 'center',
+        },
+      };
+
+      ReactDOM.render(
+        <TreeEditorExample />,
+        document.getElementById('target')
+      );
+    </script>
+  </body>
+</html>
