+++ src/component/contents/DraftEditorBlock.react.js
@@ -193,6 +193,15 @@ class DraftEditorBlock extends React.Component {
     }).toArray();
   }
 
+  _renderBlockMap(): React.Element {
+    var {getBlockChildren, block} = this.props;
+    var DraftEditorBlocks = this.props.DraftEditorBlocks;
+
+    var blocks = getBlockChildren(block.getKey());
+
+    return <DraftEditorBlocks {...this.props} blocksAsArray={blocks.valueSeq().toArray()} />;
+  }
+
   render(): React.Element<any> {
     const {direction, offsetKey} = this.props;
     const className = cx({
@@ -204,6 +213,7 @@ class DraftEditorBlock extends React.Component {
     return (
       <div data-offset-key={offsetKey} className={className}>
         {this._renderChildren()}
+        {this._renderBlockMap()}
       </div>
     );
   }
